Fix Pest test harness for CI
