+++ .github/workflows/ci.yml
@@ -14,7 +14,7 @@ jobs:
 
     steps:
     - name: Checkout code
-      uses: actions/checkout@v2
+      uses: actions/checkout@v4
 
     - name: Setup PHP
       uses: shivammathur/setup-php@v2
@@ -38,7 +38,7 @@ jobs:
       run: composer install --prefer-dist --no-progress
 
     - name: Run Lint
-      run: composer lint
+      run: composer test:lint
 
-    # - name: Run Tests -- Will add tests back after phppest issue https://github.com/pestphp/pest/issues/920 is fixed
-    #   run: composer test:unit
+    - name: Run Tests
+      run: composer test:unit
